requirement.txt
fastapi
uvicorn
python-dotenv
langchain-groq
langchain-core
langchain-community
langgraph
requests